produccion 07-05-2019

+++ app/Http/Controllers/API/UserController.php
@@ -7,6 +7,9 @@ use App\Http\Controllers\Controller;
 use App\User;
 use Illuminate\Support\Facades\Auth;
 use Validator;
+use Illuminate\Mail\Message;
+use Illuminate\Support\Facades\Mail;
+use App\Mail\RecuperarPassword;
 
 class UserController extends Controller {
 
@@ -30,8 +33,9 @@ class UserController extends Controller {
         $input['password'] = bcrypt($input['password']);
         $user = User::create($input);
         $success['token'] = $user->createToken('MyApp')->accessToken;
-        $success['nickname'] = $user->nickname;
-        return response()->json(['success' => $success], $this->successStatus);
+        $success['id'] = $user->id;
+        $success['rolId'] = $user->rol_id;
+        return response()->json($success, $this->successStatus);
     }
 
     /**
@@ -41,14 +45,19 @@ class UserController extends Controller {
      */
     public function login() {
         if (request('nickname') == null) {
-            return response()->json(['error' => 'Unauthoriseddddd'], 401);
+            return response()->json(['error' => 'Ingrese un usuario'], 401);
+        }
+        if (request('password') == null) {
+            return response()->json(['error' => 'Ingrese una contraseña'], 401);
         }
         if (Auth::attempt(['nickname' => request('nickname'), 'password' => request('password')])) {
             $user = Auth::user();
-            $success['token'] = $user->createToken('MyApp')->accessToken;
-            return response()->json(['success' => $success], $this->successStatus);
+            $success['id'] = $user->id;
+            $success['rolId'] = $user->rol_id;
+            $success['token'] = $user->createToken('MyApp')->accessToken;           
+            return response()->json($success, $this->successStatus);
         } else {
-            return response()->json(['error' => 'Unauthorisedddd'], 401);
+            return response()->json(['error' => 'Usuario o contraseña incorrecto.'], 401);
         }
     }
 
@@ -70,9 +79,9 @@ class UserController extends Controller {
      */
     public function details() {
         $user = Auth::user();
-        return response()->json(['success' => $user], $this->successStatus);
+        return User::where('id', $user->id)->get(); 
     }
-
+    
     /**
      * changePassword api
      *
@@ -84,12 +93,16 @@ class UserController extends Controller {
             return response()->json("Las contrasena no coinciden", 500);
         }
         if (isset($data['oldPassword']) && !empty($data['oldPassword']) && $data['oldPassword'] !== "" && $data['oldPassword'] !== 'undefined') {
-            $user = Auth::user();
-            $user->password = bcrypt($data['newPassword']);
-            $user->AauthAcessToken()->delete();
-            $success['token'] = $user->createToken('MyApp')->accessToken;
-            $user->save();
-            return response()->json(['success' => $success], $this->successStatus);
+            if (Auth::attempt(['nickname' => request('nickname'), 'password' => request('oldPassword')])) {
+                $user = Auth::user();
+                $user->password = bcrypt($data['newPassword']);
+                $user->AauthAcessToken()->delete();
+                $success['id'] = $user->id;
+                $success['rolId'] = $user->rol_id;
+                $success['token'] = $user->createToken('MyApp')->accessToken;   
+                $user->save();
+                return response()->json($success, $this->successStatus);
+            }
         }
         return response()->json("Error al cambiar la contrasena", 500);
     }
@@ -100,10 +113,32 @@ class UserController extends Controller {
      * @return \Illuminate\Http\Response
      */
     public function voluntario($user_id) {
-        $user = User::where('id', $user_id)->get()->toArray();
-        $user["es_voluntario"] = true;
-        $user->update();
+    //    $user = User::where('id', $user_id)->get()->toArray();
+	$user = Auth::user();
+        $user->es_voluntario = true;
+        $user->save();
         return response()->json($user, 200);
     }
 
+    public function actualizarUsuario(Request $request, User $user){
+        $user->update($request->all());
+        return response()->json($user, 200);
+    }
+
+    public function recuperarPassword(Request $request) {
+        $email = $request->input('email');
+        $email_c = $request->input('email_c');
+	if($email != $email_c){
+		return response()->json("Los correos ingresados no coinciden", 401);
+	}
+        $password = str_random(8);
+        $user = Auth::user();
+        $user->password = bcrypt($password);
+	$user->save();
+	$obj = new \stdClass();
+	$obj->nickname = $user->nickname;
+	$obj->password = $password;	
+        Mail::to($email)->send(new recuperarPassword($obj));
+        return response()->json("Se ha enviado un correo electronico con la nueva contraseña");
+    }
 }
